% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Aux_HC_CO_EF}
\alias{Aux_HC_CO_EF}
\title{CO and HC emissions factors (g/kWh) for auxiliary engines, by engine type.}
\format{
A data frame with 3 rows and 4 variables:
\describe{
\item{engineType}{auxiliary engine type, HSD=High Speed Diesel,MSD=Medium Speed Diesel,LNG=Liquid Natural Gas}
\item{co}{CO C3 marine vessel auxiliary engine emission factors (g/kWh)}
\item{hc}{Hydrocarbon C3 marine vessel auxiliary engine emission factors (g/kWh)}
\item{Source}{Character string describing source of specific emission factor for each engine type}
}
}
\usage{
data(Aux_HC_CO_EF)
}
\description{
CO and HC emissions factors (g/kWh) for auxiliary engines, by engine type.
}
\details{
This data is used in \code{calcEF_CO.r} \code{calcEF_HC.r}
}
\examples{
assign("EF",get(utils::data(Aux_HC_CO_EF)))

}
\references{
International Maritime Organization.Third IMO GHG Study 2014 - Final Report. London: International Maritime Organization, 2014.
Starcrest Consulting Group,LLC.Port of Los Angeles Inventory of Air Emissions-2016, 2017.
}
\keyword{datasets}
