% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AuxBSFC}
\alias{AuxBSFC}
\title{Brake specific fuel consumption (BSFC) (g/kWh) for auxiliary engines, by engine and fuel type.}
\format{
A data frame with 4 rows and 4 variables:
\describe{
\item{engineType}{auxiliary engine type, HSD=High Speed Diesel,MSD=Medium Speed Diesel,LNG=Liquid Natural Gas}
\item{fuelType}{fuel type used by vessel: RM/HFO=Residual Marine/Heavy Fuel Oil,MGO/MDO=Marine Gas Oil/Marine Diesel Oil,LNG=Liquid Natural Gas}
\item{BSFC}{Brake specific fuel consumption rate for auxiliary engines (g/kWh)}
\item{Source}{Character string describing source of each bsfc rate for each engine and fuel type}
}
}
\usage{
data(AuxBSFC)
}
\description{
Brake specific fuel consumption (BSFC) (g/kWh) for auxiliary engines, by engine and fuel type.
}
\details{
This data is used in \code{calcEF_CO2.r} \code{calcEF_PM.r} \code{calcEF_SO2.r}
}
\examples{
assign("EF",get(utils::data(AuxBSFC)))

}
\references{
International Maritime Organization.Third IMO GHG Study 2014 - Final Report. London: International Maritime Organization, 2014.
IVL, David Cooper, and Tomas Gustafsson. Methodology for Calculating Emissions from Ships. 1. Update of Emission Factors. n.d., 47.
Wartsila (2014). Solutions for marine and oil and gas markets.
}
\keyword{datasets}
