% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AuxNOxEF}
\alias{AuxNOxEF}
\title{NOx emission factors (g/kWh) for auxiliary engines, by fuel type, engine type, and engine tier.}
\format{
A data frame with 17 rows and 5 variables:
\describe{
\item{engineType}{auxiliary engine type, "HSD"="High Speed Diesel","MSD"="Medium Speed Diesel","LNG"="Liquid Natural Gas"}
\item{fuelType}{fuel type used by vessel: "RM/HFO"="Residual Marine/Heavy Fuel Oil","MGO/MDO"="Marine Gas Oil/Marine Diesel Oil","LNG="Liquid Natural Gas"}
\item{tier}{Engine regulatory tier (per IMO regulations). Defined by \code{calcTier.r}}
\item{nox}{NOx emission factors (g/kWh) per engine type, fuel type and tier}
\item{Source}{Character string describing source of each NOx emission factor for each engine, fuel type and tier.}
}
}
\usage{
data(AuxNOxEF)
}
\description{
NOx emission factors (g/kWh) for auxiliary engines, by fuel type, engine type, and engine tier.
}
\details{
This data is used in \code{calcEF_NOx.r}
}
\examples{
assign("EF",get(utils::data(AuxNOxEF)))

}
\references{
International Maritime Organization.Third IMO GHG Study 2014 - Final Report. London: International Maritime Organization, 2014.
Starcrest Consulting Group, LLC. San Pedro Bay Ports Emissions Inventory Methodology Report Version 2, 2021.
}
\keyword{datasets}
