% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BoilerFuelMixTable}
\alias{BoilerFuelMixTable}
\title{Fuel Mix Table for boilers, by fuel type, engine type, and location.}
\format{
A data frame with 6 rows and 4 variables:
\describe{
\item{Location}{Vessel location indicating fuel use patterns and fuel sulfur level per region. ECA indicates North American Emission Control Area}
\item{engineType}{Boiler engine type, "Boiler"}
\item{fuelType}{Fuel type used by vessel: "RM/HFO"="Residual Marine/Heavy Fuel Oil","MGO/MDO"="Marine Gas Oil/Marine Diesel Oil"}
\item{Proportion}{Indicates the proportion of vessels per engine type (from 0 to 1) which use each possible fuel within each region. Used to to calculate weighted average emission factors when vessel specific fuel information is not available. }
}
}
\usage{
data(BoilerFuelMixTable)
}
\description{
Fuel Mix Table for boilers, by fuel type, engine type, and location.
}
\details{
This data is used in \code{calcEF_CO2.r}\code{calcEF_NOx.r}\code{calcEF_PM.r}\code{calcEF_SO2.r}
}
\examples{
assign("fuelMixTable",get(utils::data(BoilerFuelMixTable)))

}
\keyword{datasets}
