% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BoilerNOxEF}
\alias{BoilerNOxEF}
\title{NOx emission factors (g/kWh) for boilers, by fuel type, engine type, and engine tier.}
\format{
A data frame with 2 rows and 5 variables:
\describe{
\item{engineType}{boiler engine type, "Boiler"}
\item{fuelType}{fuel type used by vessel: "RM/HFO"="Residual Marine/Heavy Fuel Oil","MGO/MDO"="Marine Gas Oil/Marine Diesel Oil"}
\item{tier}{Engine regulatory tier (per IMO regulations). Defined by \code{calcTier.r}}
\item{nox}{NOx emission factors (g/kWh) per engine type, fuel type and tier}
\item{Source}{Character string describing source of each NOx emission factor for each engine, fuel type and tier.}
}
}
\usage{
data(BoilerNOxEF)
}
\description{
NOx emission factors (g/kWh) for boilers, by fuel type, engine type, and engine tier.
}
\details{
This data is used in \code{calcEF_NOx.r}
}
\examples{
assign("EF",get(utils::data(BoilerNOxEF)))

}
\references{
Starcrest Consulting Group,LLC.Port of Los Angeles Inventory of Air Emissions-2016, 2017.
Starcrest Consulting Group,LLC.2014 Multi-Facility Emissions Inventory of Cargo Handling Equipment, Heavy-Duty Diesel Vehicles, Railroad Locomotives, and Commercial Marine Vessels. Starcrest Consulting Group,LLC, February 2016.
}
\keyword{datasets}
