% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEF.r
\name{calcEF}
\alias{calcEF}
\title{calcEF}
\usage{
calcEF(
  engineType,
  tier,
  location,
  loadFactor = NULL,
  ECAfuelSulfurPercentage = 0.1,
  GlobalfuelSulfurPercentage = 0.5,
  pollutants = "ALL",
  loadBasedBSFC = "N",
  output = "EF_LLAF",
  main_aux_boiler = "main",
  inputTableLocation = NULL
)
}
\arguments{
\item{engineType}{Engine type (string or vector of strings) (see
\code{\link{calcEngineType}}). Valid values are: \itemize{
\item "SSD"
\item "MSD"
\item "MSD-ED"
\item "GT"
\item "GT-ED"
\item "ST"
\item "LNG"
\item "HSD" (auxiliary only)
\item "Boiler" (boiler only)
}}

\item{tier}{NOx engine tier (string or vector of strings) (see \code{\link{calcTier}}).
Valid values are:
\itemize{
 \item "Tier 0"
 \item "Tier 1"
 \item "Tier 2"
 \item "Tier 3"
}}

\item{location}{Location of vessel (string or vector of strings). Valid values are:
\itemize{
  \item "ECA"
  \item "OutsideECA"
  \item "GreatLakes"
}}

\item{loadFactor}{Fractional percentage (between 0 and 1) of main engine
required to propel vessel at given speed (vector of numericals) (see
ShipPowerModel library). This parameter is necessary when \code{output} =
"EF_LLAF" (default) so that the correct low load adjustments are applied. It
is also necessary when \code{loadBasedBSFC} = "Y" or if "nox" is in the
pollutants list. However, if \code{output}
= "EF" \code{loadBasedBSFC} = "N", this argument will not be used and the
resulting emission factor is independent of engine load.}

\item{ECAfuelSulfurPercentage}{Fuel sulfur cap (percentage by weight) for the
Emissions Control Area (ECA). Default = 0.1\% (in effect Jan. 1, 2015)}

\item{GlobalfuelSulfurPercentage}{Fuel sulfur cap (percentage by weight) for
outside the Emissions Control Area (ECA). Default = 0.5\% (in effect Jan. 1,
2020)}

\item{pollutants}{Select pollutants (default = "ALL"). Options: \itemize{
\item "co"
\item "hc"
\item "co2"
\item "so2"
\item "pm10"
\item "pm2.5"
\item "nox"
}}

\item{loadBasedBSFC}{Determines if baseline BSFC values should be altered
according to main engine load (see \code{\link{calcLoadBSFC}}) \itemize{
\item"Y"
\item"N" (Default)
\item"lower load factor bound , upper load factor bound" (see Details)
}}

\item{output}{What output table should be produced? Options: \itemize{
\item "EF": table of corresponding emission factors
\item "EF_LLAF" (Default): table of emission factors multiplied by low load
adjustment factors
}}

\item{main_aux_boiler}{Is this calculation for a propulsive (main), auxiliary
(aux), or boiler engine? Options: \itemize{
\item "main" (Default)
\item "aux"
\item "boiler"
}}

\item{inputTableLocation}{File path (optional). Used to specify a
user-supplied set of low load adjustment factors. See details for formatting
requirements}
}
\value{
a data.table of emission factors or low load adjusted emission factors
(depending on the \code{output} argument) for the given parameters. Each
pollutant selected in the \code{pollutants} argument will be a column in the
table.
}
\description{
Creates a data table of emission factors (g/kWh) for the given
parameters.
}
\details{
Location is important for determining the fuel being used, as type of fuel
typically used varies by location.

For more information about marine emission factors, see Section 3.5 of the
Port Emissions Inventory Guidance.

If user-supplied djustment factors are used, the file should be
in .csv format with a header row. The headers should be "load" and one or more
pollutants listed in the \code{pollutants} argument. The load column should
contain numerical values in percentage space (i.e., it can range between 0 and
100). The pollutant columns should contain the specific adjustment factors for
that pollutant, corresponding to the load values. Note: The function will
linearly interpolate between the nearest values for any given load factor, and
will return an adjustment factor of 1 for load factors outside the range of
the table (i.e., it will not extrapolate).

Note: If it is desired that baseline BSFC values should be altered according
to main engine load only within a specified set of bounds, and the baseline
value should be used outside of those bounds, pass the bounds as a string
using the \code{loadBasedBSFC} argument. The bounds should be formatted as
such: the lower bound first, then a comma, then the upper bound.
}
\examples{
calcEF(engineType=c("SSD","MSD","MSD-ED","GT"),
           location=c("ECA","OutsideECA","GreatLakes","ECA"),
           loadFactor=c(0.8,0.5,0.14,0.03),
           loadBasedBSFC="0.2,1",
           output="EF_LLAF",
           pollutants="co2",
           tier=c("Tier 3","Tier 2","Tier 1", "Tier 0"))
calcEF(engineType=c("SSD","MSD","MSD-ED","GT"),
           location=c("ECA","OutsideECA","GreatLakes","ECA"),
           loadFactor=c(0.8,0.5,0.14,0.03),
           loadBasedBSFC="Y",
           output="EF_LLAF",
           main_aux_boiler="main",
           tier=c("Tier 3","Tier 2","Tier 1", "Tier 0"))
calcEF(engineType=c("SSD","SSD"),
           location=c("ECA","ECA"),
           loadFactor=c(0.8,0.03),
           output="EF",
           main_aux_boiler="main",
           tier=c("Tier 3","Tier 2"))
calcEF(engineType=c("SSD","SSD"),
           location=c("ECA","ECA"),
           loadFactor=c(0.8,0.8),
           output="EF",
           main_aux_boiler="main",
           tier=c("Tier 3","Tier 2"))
calcEF(engineType=c("HSD","MSD","LNG"),
           location=c("ECA","OutsideECA","ECA"),
           loadFactor=c(0.8,0.5,0.14),
           ECAfuelSulfurPercentage=0.1,
           GlobalfuelSulfurPercentage=2.7,
           pollutants="ALL",
           output="EF",
           main_aux_boiler="aux",
           tier=c("Tier 3","Tier 2", "Tier 0"))
calcEF(engineType=c("MSD","SSD"),
           location=c("ECA","OutsideECA"),
           loadFactor=c(0.8,0.5),
           ECAfuelSulfurPercentage=0.1,
           GlobalfuelSulfurPercentage=2.7,
           pollutants="ALL",
           output="EF",
           main_aux_boiler="boiler",
           tier=c("Tier 0", "Tier 0"))

}
\references{
\href{https://nepis.epa.gov/Exe/ZyPDF.cgi?Dockey=P10102U0.pdf}{EPA. 2020.
"Ports Emissions Inventory Guidance: Methodologies for Estimating
Port-Related and Goods Movement Mobile Source Emissions." Ann Arbor, MI:
Office of Transportation and Air Quality. US Environmental Protection Agency.}
}
