% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEF_CO.r
\name{calcEF_CO}
\alias{calcEF_CO}
\title{calcEF_CO}
\usage{
calcEF_CO(engineType, main_aux_boiler = "main")
}
\arguments{
\item{engineType}{Engine type (string or vector of strings) (see
\code{\link{calcEngineType}}). Valid values are: \itemize{
\item "SSD"
\item "MSD"
\item "MSD-ED"
\item "GT"
\item "GT-ED"
\item "ST"
\item "LNG"
\item "HSD" (auxiliary only)
\item "Boiler" (boiler only)
}}

\item{main_aux_boiler}{Is this calculation for a propulsive (main), auxiliary
(aux), or boiler engine? Options: \itemize{
\item "main" (Default)
\item "aux"
\item "boiler"
}}
}
\value{
\code{EF_CO} (g/kWh) (vector of numericals)
}
\description{
Calculates the appropriate carbon monoxide (CO) emission factor (g/kWh) for
the given parameters.
}
\details{
For more information about calculating CO emission factors, see Section 3.5.4
of the Port Emissions Inventory Guidance.
}
\examples{
calcEF_CO(c("SSD","MSD","MSD-ED"), main_aux_boiler="main")
calcEF_CO(c("HSD","MSD","LNG"), main_aux_boiler="aux")
calcEF_CO(c("Boiler"), main_aux_boiler="boiler")

}
\references{
\href{https://nepis.epa.gov/Exe/ZyPDF.cgi?Dockey=P10102U0.pdf}{EPA. 2020.
"Ports Emissions Inventory Guidance: Methodologies for Estimating
Port-Related and Goods Movement Mobile Source Emissions." Ann Arbor, MI:
Office of Transportation and Air Quality. US Environmental Protection Agency.}
}
