% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEF_CO2.r
\name{calcEF_CO2}
\alias{calcEF_CO2}
\title{calcEF_CO2}
\usage{
calcEF_CO2(engineType, location, loadFactor = NULL, main_aux_boiler = "main")
}
\arguments{
\item{engineType}{Engine type (vector of strings) (see
\code{\link{calcEngineType}}). Valid values are: \itemize{
\item "SSD"
\item "MSD"
\item "MSD-ED"
\item "GT"
\item "GT-ED"
\item "ST"
\item "LNG"
\item "HSD" (auxiliary only)
\item "Boiler" (boiler only)
}}

\item{location}{Location of vessel (vector of strings). Valid values are:
\itemize{
  \item "ECA"
  \item "OutsideECA"
  \item "GreatLakes"
}}

\item{loadFactor}{Fractional percentage (between 0 and 1) of main engine
required to propel vessel at given speed (vector of numericals) (see
ShipPowerModel library). This parameter is optional. By default, it is not
used and the resulting emission factor is independent of engine load.}

\item{main_aux_boiler}{Is this calculation for a propulsive (main), auxiliary
(aux), or boiler engine? Options: \itemize{
\item "main" (Default)
\item "aux"
\item "boiler"
}}
}
\value{
\code{EF_CO2} (g/kWh) (vector of numericals)
}
\description{
Calculates the appropriate carbon dioxide (CO2) emission factor (g/kWh) for
the given parameters.
}
\details{
Location is important for determining the fuel being used, as type of fuel
typically used varies by location.

For more information about calculating CO2 emission factors, see Section 3.5.6
of the Port Emissions Inventory Guidance.
}
\examples{
calcEF_CO2(engineType = c("SSD","MSD","LNG"),
          location = c("ECA","OutsideECA","GreatLakes"),
          loadFactor=c(0.3,.7,.9),
          main_aux_boiler = "main")

calcEF_CO2(engineType = c("SSD","SSD","LNG"),
          location = c("ECA","ECA","GreatLakes"),
          main_aux_boiler = "main")

calcEF_CO2(engineType = c("HSD","MSD","LNG"),
          location = c("ECA","ECA","GreatLakes"),main_aux_boiler="aux")

calcEF_CO2(engineType = c("MSD","SSD"),
          location = c("ECA","OutsideECA"),
          main_aux_boiler = "boiler")

}
\references{
\href{https://nepis.epa.gov/Exe/ZyPDF.cgi?Dockey=P10102U0.pdf}{EPA. 2020.
"Ports Emissions Inventory Guidance: Methodologies for Estimating
Port-Related and Goods Movement Mobile Source Emissions." Ann Arbor, MI:
Office of Transportation and Air Quality. US Environmental Protection Agency.}
}
