% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEF_HC.r
\name{calcEF_HC}
\alias{calcEF_HC}
\title{calcEF_HC}
\usage{
calcEF_HC(engineType, main_aux_boiler = "main")
}
\arguments{
\item{engineType}{Engine type (string or vector of strings) (see
\code{\link{calcEngineType}}). Valid values are: \itemize{
\item "SSD"
\item "MSD"
\item "MSD-ED"
\item "GT"
\item "GT-ED"
\item "ST"
\item "LNG"
\item "HSD" (auxiliary only)
\item "Boiler" (boiler only)
}}

\item{main_aux_boiler}{Is this calculation for a propulsive (main), auxiliary
(aux), or boiler engine? Options: \itemize{
\item "main" (Default)
\item "aux"
\item "boiler"
}}
}
\value{
\code{EF_HC} (g/kWh) (vector of numericals)
}
\description{
Calculates the appropriate hydrocarbon (HC) emission factor (g/kWh) for
the given parameters.
}
\details{
For more information about calculating HC emission factors, see Section 3.5.4
of the Port Emissions Inventory Guidance.
}
\examples{
calcEF_HC(c("SSD","MSD","MSD-ED"))
calcEF_HC(c("HSD","MSD","LNG"),main_aux_boiler="aux")
calcEF_HC(c("Boiler"), main_aux_boiler="boiler")

}
\references{
\href{https://nepis.epa.gov/Exe/ZyPDF.cgi?Dockey=P10102U0.pdf}{EPA. 2020.
"Ports Emissions Inventory Guidance: Methodologies for Estimating
Port-Related and Goods Movement Mobile Source Emissions." Ann Arbor, MI:
Office of Transportation and Air Quality. US Environmental Protection Agency.}
}
