% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEF_NOx.r
\name{calcEF_NOx}
\alias{calcEF_NOx}
\title{calcEF_NOx}
\usage{
calcEF_NOx(engineType, location, tier, main_aux_boiler = "main")
}
\arguments{
\item{engineType}{Engine type (vector of strings) (see
\code{\link{calcEngineType}}). Valid values are: \itemize{
\item "SSD"
\item "MSD"
\item "MSD-ED"
\item "GT"
\item "GT-ED"
\item "ST"
\item "LNG"
\item "HSD" (auxiliary only)
\item "Boiler" (boiler only)
}}

\item{location}{Location of vessel (vector of strings). Valid values are:
\itemize{
  \item "ECA"
  \item "OutsideECA"
  \item "GreatLakes"
}}

\item{tier}{NOx engine tier (vector of strings) (see \code{\link{calcTier}}).
Valid values are:
\itemize{
 \item "Tier 0"
 \item "Tier 1"
 \item "Tier 2"
 \item "Tier 3"
}}

\item{main_aux_boiler}{Is this calculation for a propulsive (main), auxiliary
(aux), or boiler engine? Options: \itemize{
\item "main" (Default)
\item "aux"
\item "boiler"
}}
}
\value{
\code{EF_NOx} (g/kWh) (vector of numericals)
}
\description{
Calculates the appropriate nitrogen oxide (NOx) emission factor (g/kWh) for
the given parameters.
}
\details{
Location is important for determining the fuel being used, as type of fuel
typically used varies by location.

For more information about calculating NOx emission factors, see Section 3.5.1
of the Port Emissions Inventory Guidance.
}
\examples{
calcEF_NOx(engineType = c("SSD","MSD","MSD-ED","GT"),
          location = c("ECA","OutsideECA","GreatLakes","ECA"),
          tier = c("Tier 2","Tier 3","Tier 1", "Tier 0"),
          main_aux_boiler = "main")

calcEF_NOx(engineType = c("HSD","MSD","HSD","LNG"),
          location = c("ECA","OutsideECA","GreatLakes","ECA"),
          tier = c("Tier 2","Tier 3","Tier 1", "Tier 0"),
          main_aux_boiler = "aux")

calcEF_NOx(engineType = c("Boiler","Boiler"),
          location = c("ECA","OutsideECA"),
          tier = c("Tier 2", "Tier 1"),
          main_aux_boiler = "boiler")

}
\references{
\href{https://nepis.epa.gov/Exe/ZyPDF.cgi?Dockey=P10102U0.pdf}{EPA. 2020.
"Ports Emissions Inventory Guidance: Methodologies for Estimating
Port-Related and Goods Movement Mobile Source Emissions." Ann Arbor, MI:
Office of Transportation and Air Quality. US Environmental Protection Agency.}
}
