% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEF_SO2.r
\name{calcEF_SO2}
\alias{calcEF_SO2}
\title{calcEF_SO2}
\usage{
calcEF_SO2(
  engineType,
  location,
  loadFactor = NULL,
  ECAfuelSulfurPercentage = 0.1,
  GlobalfuelSulfurPercentage = 0.5,
  main_aux_boiler = "main"
)
}
\arguments{
\item{engineType}{Engine type (vector of strings) (see
\code{\link{calcEngineType}}). Valid values are: \itemize{
\item "SSD"
\item "MSD"
\item "MSD-ED"
\item "GT"
\item "GT-ED"
\item "ST"
\item "LNG"
\item "HSD" (auxiliary only)
\item "Boiler" (boiler only)
}}

\item{location}{Location of vessel (vector of strings). Valid values are:
\itemize{
  \item "ECA"
  \item "OutsideECA"
  \item "GreatLakes"
}}

\item{loadFactor}{Fractional percentage (between 0 and 1) of main engine
required to propel vessel at given speed (vector of numericals) (see
ShipPowerModel library). This parameter is optional. By default, it is not
used and the resulting emission factor is independent of engine load.}

\item{ECAfuelSulfurPercentage}{Fuel sulfur cap (percentage by weight) for the
Emissions Control Area (ECA). Default = 0.1\% (in effect Jan. 1, 2015)}

\item{GlobalfuelSulfurPercentage}{Fuel sulfur cap (percentage by weight) for
outside the Emissions Control Area (ECA). Default = 0.5\% (in effect Jan. 1,
2020)}

\item{main_aux_boiler}{Is this calculation for a propulsive (main), auxiliary
(aux), or boiler engine? Options: \itemize{
\item "main" (Default)
\item "aux"
\item "boiler"
}}
}
\value{
\code{EF_SO2} (g/kWh) (vector of numericals)
}
\description{
Calculates the appropriate sulfur dioxide (SO2) emission factor (g/kWh) for
the given parameters.
}
\details{
Location is important for determining the fuel being used, as fuel sulfur
requirements and type of fuel typically used vary by location.

For more information about calculating SO2 emission factors, see Section 3.5.7
of the Port Emissions Inventory Guidance.
}
\examples{
calcEF_SO2(engineType = c("SSD","SSD","MSD-ED"),
          location = c("ECA","ECA","OutsideECA"),
          loadFactor = c(0.3, 0.4, 0.8))

calcEF_SO2(engineType = c("SSD","SSD","MSD-ED"),
          location = c("ECA","ECA","OutsideECA"),
          loadFactor = NULL)

calcEF_SO2(engineType = c("HSD","MSD","LNG"),
          location = c("ECA","ECA","ECA"),
          loadFactor = c(0.3, 0.4, NA),
          main_aux_boiler = "aux")

calcEF_SO2(engineType = c("Boiler","Boiler"),
          location = c("ECA","OutsideECA"),
          loadFactor = c(NA, NA, NA),
          main_aux_boiler = "boiler")

}
\references{
\href{https://nepis.epa.gov/Exe/ZyPDF.cgi?Dockey=P10102U0.pdf}{EPA. 2020.
"Ports Emissions Inventory Guidance: Methodologies for Estimating
Port-Related and Goods Movement Mobile Source Emissions." Ann Arbor, MI:
Office of Transportation and Air Quality. US Environmental Protection Agency.}
}
\seealso{
\itemize{
\item \code{\link{calcEngineType}}
\item ShipPowerModel library
}
}
