% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLLAF.r
\name{calcLLAF}
\alias{calcLLAF}
\title{calcLLAF}
\usage{
calcLLAF(
  engineType,
  location,
  loadFactor,
  ECAfuelSulfurPercentage = 0.1,
  GlobalfuelSulfurPercentage = 0.5,
  pollutants = "ALL",
  inputTableLocation = NULL
)
}
\arguments{
\item{engineType}{Engine type (vector of strings) (see
\code{\link{calcEngineType}}). Valid values are: \itemize{
\item "SSD"
\item "MSD"
\item "MSD-ED"
\item "GT"
\item "GT-ED"
\item "ST"
\item "LNG"
\item "HSD" (auxiliary only)
\item "Boiler" (boiler only)
}}

\item{location}{Location of vessel (vector of strings). Valid values are:
\itemize{
  \item "ECA"
  \item "OutsideECA"
  \item "GreatLakes"
}}

\item{loadFactor}{Fractional percentage (between 0 and 1) of main engine
required to propel vessel at given speed (vector of numericals) (see
ShipPowerModel library)}

\item{ECAfuelSulfurPercentage}{Fuel sulfur cap (percentage by weight) for the
Emissions Control Area (ECA). Default = 0.1\% (in effect Jan. 1, 2015)}

\item{GlobalfuelSulfurPercentage}{Fuel sulfur cap (percentage by weight) for
outside the Emissions Control Area (ECA). Default = 0.5\% (in effect Jan. 1,
2020)}

\item{pollutants}{Select pollutants (default = "ALL"). Options: \itemize{
\item "co"
\item "hc"
\item "co2"
\item "so2"
\item "pm10"
\item "pm2.5"
\item "nox"
}}

\item{inputTableLocation}{File path (optional). Used to specify a
user-supplied set of adjustment factors. See details for formatting
requirements}
}
\value{
a data.table of low load adjusted emission factors the given parameters. Each
pollutant selected in the \code{pollutants} argument will be a column in the
table.
}
\description{
Creates a data table of low load adjustment factors (unitless)
for the given parameters.
}
\details{
Location is important for determining the fuel being used, as type of fuel
typically used varies by location. This impacts the calculation of LLAF
for SO2.

For more information about the default low load adjustment factors, see
Section 3.7 of the Port Emissions Inventory Guidance.

If user-supplied adjustment factors are used, the file should be
in .csv format with a header row. The headers should be "load" and one or more
pollutants listed in the \code{pollutants} argument. The load column should
contain numerical values in percentage space (i.e., it can range between 0 and
100). The pollutant columns should contain the specific adjustment factors for
that pollutant, corresponding to the load values. Note: This function will
linearly interpolate between the nearest values for any given load factor, and
will return an adjustment factor of 1 for load factors outside the range of
the table (i.e., it will not extrapolate).
}
\examples{
calcLLAF(engineType=c("SSD","MSD","MSD-ED","GT"),
        location = "ECA",
        loadFactor=c(0.8,0.5,0.14,0.03),
        pollutants="co2")
calcLLAF(engineType=c("SSD","MSD","MSD-ED","GT"),
        location = c("ECA","GreatLakes","ECA","OutsideECA"),
        loadFactor=c(0.8,0.5,0.14,0.03))
calcLLAF(engineType=c("HSD","MSD","LNG"),
        location = c("ECA","GreatLakes","ECA","OutsideECA"),
        loadFactor=c(0.8,0.5,0.14),
        ECAfuelSulfurPercentage=0.1,
        GlobalfuelSulfurPercentage=2.7,
        pollutants="ALL")

}
\references{
\href{https://nepis.epa.gov/Exe/ZyPDF.cgi?Dockey=P10102U0.pdf}{EPA. 2020.
"Ports Emissions Inventory Guidance: Methodologies for Estimating
Port-Related and Goods Movement Mobile Source Emissions." Ann Arbor, MI:
Office of Transportation and Air Quality. US Environmental Protection Agency.}
}
