% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLoadBSFC.r
\name{calcLoadBSFC}
\alias{calcLoadBSFC}
\title{calcLoadBSFC}
\usage{
calcLoadBSFC(loadFactor = NULL, BSFC_Baseline)
}
\arguments{
\item{loadFactor}{Fraction of total installed propulsive power representing
the instantaneous or average propulsive power (unitless) (vector of
numericals) (see ShipPowerModel library). If \code{NA} is supplied, the
baseline BSFC value is returned (i.e., the resulting emission factor will be
independent of engine load).}

\item{BSFC_Baseline}{Baseline BSFC value (g/kWh) (vector of numericals)}
}
\value{
\code{BSFC} (g/kWh) (vector of numericals)
}
\description{
Calculates Brake Specific Fuel Consumption (BSFC) or Specific Oil
Fuel Consumption (SFOC) using a parabolic relationship with the main engine
load factor.
}
\examples{
calcLoadBSFC(c(NA, 0.5, 0.7),c(185,195,300))

}
\references{
International Maritime Organization. 2014. "Third IMO GHG study 2014 - Final
report." London: International Maritime Organization.

\href{https://doi.org/10.5194/acp-12-2641-2012}{Jalkanen, J.-P., Johansson
L., Kukkonen, J., Brink, A., Kalli, J., and Stipa , T. 2012. "Extension of an
assessment model of ship traffic exhaust emissions for particulate matter and
carbon monoxide. Atmospheric Chemistry and Physics 12, no. 5.}
}
