% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcShipCategory.r
\name{calcShipCategory}
\alias{calcShipCategory}
\title{calcShipCategory}
\usage{
calcShipCategory(mainEngineBore, mainEngineStroke)
}
\arguments{
\item{mainEngineBore}{Diameter of the main engine cylinder (mm) (vector of
numericals)}

\item{mainEngineStroke}{Main engine stroke length, the distance travelled by
the piston in each cycle (mm) (vector of numericals)}
}
\value{
\code{shipCategory} (vector of ints). Valid values are: \itemize{
\item 1
\item 2
\item 3
}
}
\description{
Assigns EPA regulatory ship category from engine displacement
}
\details{
For more information, see Section 3.3.2.1 of the Port Emissions Inventory
Guidance.
}
\examples{
calcShipCategory(400, 240)

}
\references{
\href{https://nepis.epa.gov/Exe/ZyPURL.cgi?Dockey=P1005ZGH.TXT}{EPA. 2009.
"Regulatory impact analysis: Control of emissions air pollution from category
3 marine diesel engines." Ann Arbor, MI: Office of Transportation and Air
Quality. US Environmental Protection Agency.}

\href{https://nepis.epa.gov/Exe/ZyPDF.cgi/?Dockey=P10024CN.PDF}{EPA. 2008.
"Regulatory impact analysis: Control of emissions of air pollution from
locomotive engines and marine compression ignition engines less than 30 liters
per cylinder." Ann Arbor, MI: Office of Transportation and Air
Quality. US Environmental Protection Agency.}

\href{https://nepis.epa.gov/Exe/ZyPDF.cgi?Dockey=P10102U0.pdf}{EPA. 2020.
"Port Emissions Inventory Guidance: Methodologies for Estimating
Port-Related and Goods Movement Mobile Source Emissions." Ann Arbor, MI:
Office of Transportation and Air Quality. US Environmental Protection Agency.}
}
