% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTestIMO.r
\name{calcTestIMO}
\alias{calcTestIMO}
\title{calcTestIMO}
\usage{
calcTestIMO(IMO, correctFlag = "Correct", incorrectFlag = "Incorrect")
}
\arguments{
\item{IMO}{IMO number (vector of numericals or strings)}

\item{correctFlag}{Output indicator for rows where the IMO Number is in the
correct format}

\item{incorrectFlag}{Output indicator for rows where the IMO Number is not in
the correct format}
}
\value{
\code{IMO_number_flag}, a factorized vector containing either the value of
\code{correctFlag} or \code{incorrectFlag}, depending on the correctness of
the IMO number
}
\description{
Determines whether or not an IMO number is in the correct format. IMO numbers
should have 7 digits with a check digit at the end.
}
\examples{
calcTestIMO(IMO=c(899046900,0,9394935))

}
