% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CO2ConversionFactor}
\alias{CO2ConversionFactor}
\title{Conversion factors to produce CO2 emission factors from BSFC values by fuel type.}
\format{
A data frame with 3 rows and 3 variables:
\describe{
\item{fuelType}{Fuel type used by vessel: "RM/HFO"="Residual Marine/Heavy Fuel Oil","MGO/MDO"="Marine Gas Oil/Marine Diesel Oil","LNG="Liquid Natural Gas"}
\item{conversionFactor}{Conversion factors to convert BSFC values to CO2 emission factors.}
\item{Source}{Character string describing source of each CO2 conversion factor by fuel type.}
}
}
\usage{
data(CO2ConversionFactor)
}
\description{
Conversion factors to produce CO2 emission factors from BSFC values by fuel type.
}
\details{
This data is used in \code{calcEF_CO2.r}
}
\examples{
  data(CO2ConversionFactor)

}
\references{
Marine Environmental Protection Committee. 2012 Guidelines on the Method of Calculation of the Attained Energy Efficiency Design Index (EEDI) for New Ships (MEPC 63/28 Annex 8), March 2012.
International Maritime Organization.Third IMO GHG Study 2014 - Final Report. London: International Maritime Organization, 2014.
}
\keyword{datasets}
