% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EF_PM_ST_GT_LNG}
\alias{EF_PM_ST_GT_LNG}
\title{PM10 emission factors (g/kWh) for non diesel engines per fuel type and engine type.}
\format{
A data frame with 7 rows and 4 variables:
\describe{
\item{engineType}{Non-diesel engine type: "ST"="Steam Turbine","GT"="Gas Turbine","GT-ED"="Gas Turbine Electic Diesel","LNG"="Liquid Natural Gas"}
\item{fuelType}{Fuel type used by vessel: "RM/HFO"="Residual Marine/Heavy Fuel Oil","MGO/MDO"="Marine Gas Oil/Marine Diesel Oil","LNG="Liquid Natural Gas"}
\item{pm10EF}{PM10 emission factors (g/kWh) per engine type and fuel type.}
\item{Source}{Character string describing source of each PM emission factors by fuel and engine type.}
}
}
\usage{
data(EF_PM_ST_GT_LNG)
}
\description{
PM10 emission factors (g/kWh) for non diesel engines per fuel type and engine type.
}
\details{
This data is used in \code{calcEF_PM.r}
}
\examples{
data(EF_PM_ST_GT_LNG)
}
\references{
International Maritime Organization.Third IMO GHG Study 2014 - Final Report. London: International Maritime Organization, 2014.
IVL, David Cooper, and Tomas Gustafsson. Methodology for Calculating Emissions from Ships. 1. Update of Emission Factors, n.d., 47.
Kristensen, Hans Otto. Energy Demand and Exhaust Gas Emissions of Marine Engines, 2012, 30.
}
\keyword{datasets}
