% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EFSulfurEQCoefficients}
\alias{EFSulfurEQCoefficients}
\title{Coefficients required to calculate PM emission factors (g/kWh) per fuel type for diesel engines.}
\format{
A data frame with 2 rows and 5 variables:
\describe{
  \item{fuelType}{Fuel type used by vessel: "RM/HFO"="Residual Marine/Heavy Fuel Oil","MGO/MDO"="Marine Gas Oil/Marine Diesel Oil"}
  \item{MWR}{Molecular weight ratio of sulfate PM to sulfur (224/32=7).}
  \item{FSC}{Percantage of sulfur in fuel that is converted to direct sulfate PM.}
  \item{Snom}{Nominal fuel sulfur level.}
  \item{PMnom}{Base PM emission rate for a fuel sulfur level of 0\%.}
  \item{Source}{Character string describing source of each CO2 conversion factor by fuel type.}
}
}
\usage{
data(EFSulfurEQCoefficients)
}
\description{
Coefficients required to calculate PM emission factors (g/kWh) per fuel type for diesel engines.
}
\details{
This data is used in \code{calcEF_PM.r}
}
\examples{
data(EFSulfurEQCoefficients)
}
\references{
International Maritime Organization.Third IMO GHG Study 2014 - Final Report. London: International Maritime Organization, 2014.
EPA. Regulatory Impact Analysis: Control of Emissions of Air Pollution from Category 3 Marine Deisel Engines, 2009.
}
\keyword{datasets}
